--- a/src/main/java/org/openstreetmap/atlas/checks/validation/tag/RoadNameSpellingConsistencyCheck.java
+++ b/src/main/java/org/openstreetmap/atlas/checks/validation/tag/RoadNameSpellingConsistencyCheck.java
@@ -76,11 +76,10 @@ public class RoadNameSpellingConsistencyCheck extends BaseCheck<Long>
         // Collect all of the Edges within a search distance, then filter those that have NameTags
         // that are slightly different than edge
         final Set<Edge> inconsistentEdgeSet = new RoadNameSpellingConsistencyCheckWalker(edge,
-                this.maximumSearchDistance)
-                        .collectEdges().stream()
+                this.maximumSearchDistance).collectEdges().stream().parallel()
+                        .filter(incomingEdge -> !this.isFlagged(incomingEdge.getIdentifier()))
                         .filter(RoadNameSpellingConsistencyCheckWalker
                                 .isEdgeWithInconsistentSpelling(edge))
-                        .filter(incomingEdge -> !this.isFlagged(incomingEdge.getIdentifier()))
                         .collect(Collectors.toSet());
 
         // If the Walker found any inconsistent NameTag spellings