--- a/src/main/java/org/openstreetmap/atlas/checks/validation/tag/RoadNameSpellingConsistencyCheck.java
+++ b/src/main/java/org/openstreetmap/atlas/checks/validation/tag/RoadNameSpellingConsistencyCheck.java
@@ -76,7 +76,7 @@ public class RoadNameSpellingConsistencyCheck extends BaseCheck<Long>
         // Collect all of the Edges within a search distance, then filter those that have NameTags
         // that are slightly different than edge
         final Set<Edge> inconsistentEdgeSet = new RoadNameSpellingConsistencyCheckWalker(edge,
-                this.maximumSearchDistance).collectEdges().stream().parallel()
+                this.maximumSearchDistance).collectEdges().stream()
                         .filter(incomingEdge -> !this.isFlagged(incomingEdge.getIdentifier()))
                         .filter(RoadNameSpellingConsistencyCheckWalker
                                 .isEdgeWithInconsistentSpelling(edge))