--- a/src/main/java/org/openstreetmap/atlas/checks/validation/tag/RoadNameSpellingConsistencyCheckWalker.java
+++ b/src/main/java/org/openstreetmap/atlas/checks/validation/tag/RoadNameSpellingConsistencyCheckWalker.java
@@ -209,18 +209,17 @@ class RoadNameSpellingConsistencyCheckWalker extends EdgeWalker
         incomingEdgeNameAlphanumericIdentifierStrings = Arrays
                 .stream(incomingEdgeName.split(WHITESPACE_REGEX))
                 .filter(substring -> Stream.of(CJKNumbers.values())
-                                .anyMatch(cjkNumber -> substring.contains(
-                                        new String(Character.toChars(cjkNumber.getValue()))))
+                        .anyMatch(cjkNumber -> substring
+                                .contains(new String(Character.toChars(cjkNumber.getValue()))))
                         || ALPHANUMERIC_IDENTIFIER_STRING_PATTERN.matcher(substring).matches()
-                        || CHARACTER_IDENTIFIER_STRING_PATTERN.matcher(substring).matches()
-                        )
+                        || CHARACTER_IDENTIFIER_STRING_PATTERN.matcher(substring).matches())
                 .collect(Collectors.toList());
 
         final List<String> startingEdgeNameAlphanumericIdentifierStrings = Arrays
                 .stream(startingEdgeName.split(WHITESPACE_REGEX))
                 .filter(substring -> Stream.of(CJKNumbers.values())
-                        .anyMatch(cjkNumber -> substring.contains(
-                                new String(Character.toChars(cjkNumber.getValue()))))
+                        .anyMatch(cjkNumber -> substring
+                                .contains(new String(Character.toChars(cjkNumber.getValue()))))
                         || ALPHANUMERIC_IDENTIFIER_STRING_PATTERN.matcher(substring).matches()
                         || CHARACTER_IDENTIFIER_STRING_PATTERN.matcher(substring).matches())
                 .collect(Collectors.toList());