### Description:

This PR is created to improve the performance of the RoadNameSpellingConsistencyCheck <rdar://problem/57118015>

Myself and Sean performed some tests on smaller 348MB and on larger data set. We came to a conclusion that regex used in RoadNameSpellingConsistencyCheckWalker.java is consuming considerable amount of time to run the check. Used profiler in IntelliJ to understand the metrics.

Handling the logic without regex is complex. So, made few minor changes to the existing code base to reduce the time to at least few seconds.

### Potential Impact:
Execution time before changes are 78 to 82 secs for 340MB test data.
Execution time after changes are 70 to 74 secs for 340 MB test data.

### Unit Test Approach:
No changes for the tests.

### Test Results:
Profiler output on 340 MB test data is as follows:
<img width="1635" alt="Profiler_output" src="https://user-images.githubusercontent.com/16053580/70178841-92b3ee80-16a2-11ea-87b9-b67123c9483c.png">